requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.1.1
django-ckeditor==6.5.1
django-cors-headers==3.13.0
django-js-asset==2.0.0
django-taggit==3.0.0
django-taggit-serializer==0.1.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.0
Pillow==9.2.0
PyJWT==2.5.0
pytz==2022.2.1
six==1.16.0
sqlparse==0.4.3
